+++ CHANGES.rst
@@ -32,6 +32,7 @@ v2.0.0 (xxxx-xx-xx)
 - Modify ``chain()`` to be lazy and only compute the final value when ``value()`` called.
 - Modify ``flatten()`` to be shallow by default, remove callback option, and add ``is_deep`` option. (**breaking change**)
 - Modify ``is_number()`` to return ``False`` for boolean ``True`` and ``False``. (**breaking change**)
+- Move predicate functions from ``pydash.api.objects`` to ``pydash.api.predicates``. (**possible breaking change**)
 - Rename ``create_callback()`` to ``iteratee``. (**breaking change**)
 - Rename *private* utility function ``_iter_callback()`` to ``itercallback()``. (**breaking change**)
 - Rename *private* utility function ``_iter_list_callback()`` to ``iterlist_callback()``. (**breaking change**)